sync cart count with localstorage
